src/main/webapp/app/entities/basketball-court/delete/basketball-court-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { provideHttpClient, HttpResponse } from '@angular/common/http';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { BasketballCourtService } from '../service/basketball-court.service';

import { BasketballCourtDeleteDialogComponent } from './basketball-court-delete-dialog.component';

describe('BasketballCourt Management Delete Component', () => {
  let comp: BasketballCourtDeleteDialogComponent;
  let fixture: ComponentFixture<BasketballCourtDeleteDialogComponent>;
  let service: BasketballCourtService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [BasketballCourtDeleteDialogComponent],
      providers: [provideHttpClient(), NgbActiveModal],
    })
      .overrideTemplate(BasketballCourtDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(BasketballCourtDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(BasketballCourtService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
